rename: chip fix
