+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as Gnb } from "./gnb/gnb";
 export { default as Header } from "./header/header";
 export { default as DropdownMenu } from "./dropdown-menu/dropdown-menu";
-export { default as Chip } from "./chip/Chip";
+export { default as Chip } from "./chip/chip";
 export { default as Flavor } from "./flavor/flavor";
 export { default as TextInput } from "./text-input/text-input";
 export { default as SelectType } from "./select-type/select-type";
